+++ config/http/sessions.php
@@ -1,4 +1,8 @@
 <?php
+use Opulence\Cache\FileBridge;
+use Opulence\Framework\Configuration\Config;
+use Opulence\Sessions\Handlers\FileSessionHandler;
+
 /**
  * ----------------------------------------------------------
  * Define the session config
@@ -10,10 +14,12 @@ return [
      * General settings
      * ----------------------------------------------------------
      *
+     * "handler" => The name of the session handler class
      * "lifetime" => Lifetime of the session in seconds
      * "name" => The name of the session
      * "isEncrypted" => Whether or not the session data is encrypted when stored
      */
+    "handler" => $environment->getVar("SESSION_HANDLER", FileSessionHandler::class),
     "lifetime" => 7200,
     "name" => "__opulence_session",
     "isEncrypted" => false,
@@ -37,10 +43,12 @@ return [
      * These are useful if you use a cookie to remember the session Id between requests
      *
      * "cookie.domain" => The domain of the cookie
+     * "cookie.isHttpOnly" => Whether or not the cookie is HTTP-only
      * "cookie.isSecure" => Whether or not the cookie is secure
      * "cookie.path" => The path of the cookie
      */
     "cookie.domain" => $environment->getVar("SESSION_COOKIE_DOMAIN", ""),
+    "cookie.isHttpOnly" => true,
     "cookie.isSecure" => $environment->getVar("SESSION_COOKIE_IS_SECURE", false),
     "cookie.path" => $environment->getVar("SESSION_COOKIE_PATH", "/"),
 
@@ -49,9 +57,11 @@ return [
      * Cache-backed session settings
      * ----------------------------------------------------------
      *
+     * "cache.bridge" => The name of the cache bridge class your sessions use
      * "cache.clientName" => The name of the client to use in your cache bridge
      * "cache.keyPrefix" => The prefix to use on all cache keys to avoid naming collisions
      */
+    "cache.bridge" => $environment->getVar("SESSION_CACHE_BRIDGE", FileBridge::class),
     "cache.clientName" => "default",
     "cache.keyPrefix" => "opulence:",
 
@@ -62,7 +72,7 @@ return [
      *
      * "file.path" => The path of the session file
      */
-    "file.path" => realpath(__DIR__ . "/../../tmp/framework/http/sessions"),
+    "file.path" => Config::get("paths", "tmp.framework.http") . "/sessions",
 
     /**
      * ----------------------------------------------------------